# src/Web.Client/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/completed-products/on-storage": {
      "get": {
        "tags": [
          "CompletedProduct"
        ],
        "summary": "Список всех готовых продуктов на складе",
        "operationId": "GetAllCompletedProductsOnStorage",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 1
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            },
            "x-position": 2
          },
          {
            "name": "filterExpression",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "filterValues",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string",
                "nullable": true
              }
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/completed-products/product/{productId}": {
      "get": {
        "tags": [
          "CompletedProduct"
        ],
        "summary": "Список готовых продуктов по ProductID (справочный ID)",
        "operationId": "GetAllProductsOnStorageByProductId",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompletedProductModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/manufacturing-process/{manufacturingProcessId}": {
      "get": {
        "tags": [
          "Issue"
        ],
        "summary": "Получение всех задач производственного процесса",
        "operationId": "GetIssuesByManufacturingProcessId",
        "parameters": [
          {
            "name": "manufacturingProcessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IssueFromManufacturingProcessModel"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/manufacturingProcesses": {
      "post": {
        "tags": [
          "ManufacturingProcess"
        ],
        "summary": "Планирование производственного процесса",
        "operationId": "ScheduleManufacturingProcess",
        "requestBody": {
          "x-name": "command",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledManufacturingProcessCommand"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ManufacturingProcess"
        ],
        "summary": "Получить все производственные процессы",
        "operationId": "GetAllManufacturingProcesses",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 1
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            },
            "x-position": 2
          },
          {
            "name": "filterExpression",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "filterValues",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string",
                "nullable": true
              }
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManufacturingProcessResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manufacturingProcesses/{id}": {
      "get": {
        "tags": [
          "ManufacturingProcess"
        ],
        "operationId": "GetManufacturingProcessById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturingProcessWithDetailsModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manufacturingProcesses/{id}/launch": {
      "put": {
        "tags": [
          "ManufacturingProcess"
        ],
        "summary": "Запуск производственного процесса",
        "operationId": "LaunchManufacturingProcess",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manufacturingProcesses/{id}/complete": {
      "put": {
        "tags": [
          "ManufacturingProcess"
        ],
        "summary": "Завершение производственного процесса",
        "operationId": "CompleteManufacturingProcess",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/nofitications-for-users": {
      "get": {
        "tags": [
          "NotificationForUsers"
        ],
        "summary": "Получить все уведомления для пользователей",
        "operationId": "GetAllNotifications",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 1
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationForUsersModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "NotificationForUsers"
        ],
        "summary": "Добавить уведомление для пользователей",
        "operationId": "PublishNotification",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationForUsersRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Получение продукта по id",
        "operationId": "GetProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "Удаление продукта по id",
        "operationId": "DeleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompletedProductModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "productDevelopmentId": {
            "type": "string",
            "format": "guid"
          },
          "serialNumber": {
            "type": "string"
          }
        }
      },
      "IssueFromManufacturingProcessModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "operationName": {
            "type": "string"
          },
          "issueNumber": {
            "type": "integer",
            "format": "int32"
          },
          "executorId": {
            "type": "string",
            "format": "guid"
          },
          "responsibleId": {
            "type": "string",
            "format": "guid"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "currentQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatusModel"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "productCompositionFlag": {
            "type": "boolean"
          }
        }
      },
      "IssueStatusModel": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Pending",
          "PendingPreviousIssue",
          "InProgress",
          "Completed",
          "Cancelled"
        ],
        "enum": [
          100,
          101,
          200,
          300,
          400
        ]
      },
...
      "ScheduledManufacturingProcessCommand": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "technologicalCardId": {
            "type": "string",
            "format": "guid"
          },
          "batchNumber": {
            "type": "string"
          },
          "targetQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduledIssue"
            }
          }
        }
      },
      "ScheduledIssue": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "stepId": {
            "type": "string",
            "format": "guid"
          },
          "responsibleId": {
            "type": "string",
            "format": "guid"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "productCompositionSettingFlag": {
            "type": "boolean"
          }
        }
      },
      "ManufacturingProcessWithDetailsModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "productId": {
            "type": "string",
            "format": "guid"
          },
          "productName": {
            "type": "string"
          },
          "techCardId": {
            "type": "string",
            "format": "guid"
          },
          "techCardNumber": {
            "type": "string"
          },
          "batchNumber": {
            "type": "string"
          },
          "currentQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "targetQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ManufacturingProcessStatusModel"
          }
        }
      },
      "ManufacturingProcessStatusModel": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Pending",
          "InPaused",
          "InProgress",
          "Completed",
          "Cancelled"
        ],
        "enum": [
          100,
          101,
          200,
          300,
          400
        ]
      },
      "ManufacturingProcessResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "productId": {
            "type": "string",
            "format": "guid"
          },
          "productName": {
            "type": "string"
          },
          "techCardId": {
            "type": "string",
            "format": "guid"
          },
          "techCardNumber": {
            "type": "string"
          },
          "batchNumber": {
            "type": "string",
            "description": "Номер партии"
          },
          "currentQuantity": {
            "type": "integer",
            "description": "Текущее количество продукта из производственного процесса",
            "format": "int32"
          },
          "targetQuantity": {
            "type": "integer",
            "description": "Целевое количество продукта в производственном процессе",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "description": "Дата начала выполнения производственного процесса",
            "format": "date-time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "description": "Дата завершения выполнения производственного процесса",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ManufacturingProcessStatus"
          }
        }
      },
      "ManufacturingProcessStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Pending",
          "InPaused",
          "InProgress",
          "Completed",
          "Cancelled"
        ],
        "enum": [
          100,
          101,
          200,
          300,
          400
        ]
      },
...
      "NotificationForUsersModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "message": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "NotificationForUsersRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ProductResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "format": "guid"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedById": {
            "type": "string",
            "format": "guid"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletedById": {
            "type": "string",
            "format": "guid",
            "nullable": true
          }
        }
      },
...
    }
  }
}